backend/requirements.txt
fastapi
uvicorn
websockets
opencv-python
ultralytics
numpy
fpdf2
easyocr
